pyproject.toml
[tool.poetry]
name = "rootshell_platform_api"
version = "0.2.0"
description = "A Package to interact with Rootshells Public API"
authors = ["Rootshell Developers <chris.mills@rootshellsecurity.net>"]
readme = "README_PI.md"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.32.3"
python-dotenv = "^1.0.1"

click = "^8.1.7"
pytest = "^8.2.2"
[tool.poetry.group.dev.dependencies]
requests-mock = "^1.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
rootshell_platform = "rootshell_platform_api.cli:cli"
